rust: add -lresolv to fix MacOS tests
Summary:
Rust's libstd has a stray dependency on libresolv. Normally this is satisfied when rustc is doing
the linking, but when generating standalone Rust libraries for use as a C++ dependency it still
has the libresolv requirement. For now, just add `-lresolv` to the link line in the affected test.

This is a bit of a hack; it shouldn't require libresolv at all.
Upstream rust issue https://github.com/rust-lang/rust/issues/46797 is tracking this.

Test Plan: unit tests

Reviewed By: styurin

fbshipit-source-id: 734b1dc