Fix blacklisting SwiftRuntimeNativeLinkable
Summary:
Tests with test host and swift in both the test host and test target currently fail compilation. Because the runtime is a dependency of the bundle_loader, it is blacklisted and not linked in the test target. However, if there's usage of swift in the tests their object files will contain linker directives to link against the standard library. Need to validate if there's a difference between static and dynamic linking, as only macos has static linking and ios does not.