Add host_info() method to get os and cpu architecture for host
Summary: Adds a host_info() method that allows us to not call out to system.platform()/system.machine() as many people do. This includes both the python and skylark definition. This will likely be deprecated once full platform support makes it in, but it's a reasonable stopgap for the moment as we migrate some stuff to skylark.

Test Plan: Added unittests, built a project that printed out the host info (both python and skylark), and verified docs in soyweb-local.sh

Reviewed By: styurin

fbshipit-source-id: 7808023