Remove the 'command' property from the compilation database
Summary:
Some dbs are huge, like this one from the Messenger app /Users/wallace/fbsource/fbobjc/buck-out/gen/Apps/FBMessenger/__MNMessagingApp#compilation-database/compile_commands.json, which uses 2GB of disk space.
And there are a lot of them and hinder the language services that can be provided by IDEs like Nuclide. In order to improve this situation, I'm removing the 'command' property from the compilation databases, leaving only 'arguments'.

Test Plan: After the change, the improvement in the generation time for that db was around 5 secs and it uses 1GB of disk.

Reviewed By: dinhviethoa

fbshipit-source-id: f22a351