Always pass -d flag to install when possible
Summary:
By default, `pm install` refuses to downgrade apps on Android unless you pass the `-d` flag, which triggers `PackageManager.INSTALL_ALLOW_DOWNGRADE`.

Buck doesn't have an easy way to pass the `-d` flag, and even if it did it would be annoying to train everyone to use it all the time. Just pass `-d` always, which seems like a sensible default.

Test Plan: Trigger a downgrade, and verify that it succeeds.

Reviewed By: dreiss

fbshipit-source-id: 9cb94f7