Use consistent way to construct attribute keys.
Summary:
Unfortunately attribute map uses strings as keys (something
that I would love to fix in the future) instead of `TargetNode`s, which
can cause bugs if attempt to access them using a different key is
made. To prevent this, I have introduced this a common shared method,
but forgot to update one of the call-sites.

Test Plan: CI

Reviewed By: stupaq-fb

fbshipit-source-id: 6a61f06