Cache rule keys of common dependencies between pipelined rules
Summary:
Java library rules have the exact same set of dependencies as their `#source-abi` rules, except
they also depend on the `#source-abi` rules themselves. By default, these are different sets
of dependencies which must be hashed independently. This diff makes the relationship explicit
and encodes it in the rule key as such, so that the library rulekeys can get cache hits for
the common dependencies.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: c45df43