Print out when querying for owners of files that do not exist
Summary: Add some logging when files do not exist / are not owned. This is to alleviate user errors where an incorrect filename was provided. It was hard to tell whether there was no owner, or whether an incorrect path was given.

Test Plan: Updated tests, ran gainst a sample project where files didn't exist or weren't owned, and I got error messages in both cases on stderr.

Reviewed By: sbalabanov

fbshipit-source-id: eed2ce0