Allow duplicate keys in minrank query.
Summary:
Flavored nodes with same attributes map into the same
keys because their flavor is removed to form a key. This is a
questionable decision, but is what unfortunately was already
provided by Buck :(, so until we decide that it's safe to start
using flavored names instead of unflavored, duplicates are allowed
by switching to a regular Java map that allows duplicates (just like
it used to).

Test Plan:
CI
unfortunately I could not come up with a repro test case :(

Reviewed By: jtorkkola

fbshipit-source-id: 660ef34