Remove "out" from AbstractGenruleDescription.CommonArg
Summary:
Refactor `AbstractGenruleDescription` so that its subclasses can
choose whether they want to support the `out` argument or not.
While working on this, I noticed that `apk_genrule` does not use
the `out` argument anywhere, so I updated its documentation and
made that argument optional so that it can be removed in the future.

Test Plan: Unit tests.

Reviewed By: philipjameson

fbshipit-source-id: abf51c1