Implement jar_genrule
Summary:
Implement a new rule, `jar_genrule`, that mostly behaves the same
as a regular `genrule`, but its output is treated similarly to
the output of a `java_binary`, meaning that it can be executed with
`buck run` and it can be used with the `$(exe )` macro in other
genrules. The aim of the new rule is to allow to post-process Java
binaries without affecting code that already uses them.

Test Plan:
Added an integration test.
Manual testing with existing Java binaries.

Reviewed By: philipjameson

fbshipit-source-id: 1a755b2