Fix `platform_exe` parameter of `command_alias`
Summary:
Adapts the implementation of `CommandAlias` so that not only the rule key of the build rule itself is stable regardless of the host platform, but also the rule key of the `Tool` returned by `CommandAlias#getExecutableCommand()`.

To make that work, `Tool` instances are created for every platform, and used for rule key calculations.

Test Plan: - extended and reorganized unit and integration tests for `CommandAlias` / `CommandAliasDescription`

Reviewed By: dinhviethoa

fbshipit-source-id: 76acb22