Make DistBuild create one RuleKeyCacheScope
Summary:
Previously, we pass around RuleKeyCache as an Optional, and create the actual RuleKeyCache
in several different locations.
This makes the code hard to follow, and difficult to instrument the cache.
This change refactors creation of RuleKeyCache to a RuleKeyCacheScope at top level, and removes the optional.

Test Plan: ran all distributed build tests

Reviewed By: sbalabanov

fbshipit-source-id: c6dab93