Add test for race condition with parallel action graph construction
Summary:
ForkJoinPool is not always smart enough to deal with dependent tasks and may deadlock
if all it's thread are running tasks A...D created by `submit()` and the running tasks are blocked by
a task E create through `fork()`. ForkJoinPool will not switch out of the blocked threads to run task E.
If however, A..D are created with `fork()`, ForkJoinPool will be able to switch out of blocked threads
to run E.

This test should catch problems if we change our `submit()` and `fork()` usages, or if ForkJoinPool behaviour
changes.

Test Plan: unit test

Reviewed By: yiding, sbalabanov

fbshipit-source-id: 79b6e66