Link libatomic when libc++abi and ARMEABI is used
Summary:
When ARMEABI is used, libc++abi requires libatomic. This makes the following command fail:
  buck build mode/opt //libraries/fresco/imagepipeline/src/main/jni:jni_static#android-arm,shared --show-output buck-configs/buckconfig.fblite_arm
with undefined refrences to '__atomic_exchange_4'.
Adding a flag to loader_flags isn't enough, because we need "-latomic" to be added AFTER "-lc++abi", and it's the last linker flag (and also, as can be seen here, "-lc++abi" isn't always added).

Test Plan:
Compile buck and see that this succeeds:
  buck build mode/opt //libraries/fresco/imagepipeline/src/main/jni:jni_static#android-arm,shared --show-output buck-configs/buckconfig.fblite_arm
Also see that these still work:
  buck build mode/opt //libraries/fresco/imagepipeline/src/main/jni:jni_static#android-x86,shared --show-output buck-configs/buckconfig.fblite_arm
  buck build mode/opt //libraries/fresco/imagepipeline/src/main/jni:jni_static#android-armv7,shared --show-output buck-configs/buckconfig.fblite_arm

fbshipit-source-id: 7a98f36