python: don't use `PackagedResource` for `__test_main__.py`
Summary:
The `PackagedResource` class is somewhat deprecated and leads to unfriendly
user errors if `buck-out` is removed without killing the daemon (which is a
common flow).  This diff sets up a per-binary rule to unpack the test main,
which should also allow subsequent changes to override the test main on per-
binary basis.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 467dbc7