Fixed shutdown bug preventing local build failing if remote build failed
Summary:
CountDownLatch was being initialized with value 10 by mistake, which meant that if distributed build fails we cannot proceed to kill the local build.

It is now initialized with value 1, which means we can proceed with the cancellation.

Additionally changed if statement so that shouldAlwaysWaitForRemoteBuildBeforeProceedingLocally is ignored.

Test Plan: ci

Reviewed By: shivanker

fbshipit-source-id: 14b5ded