Stop dumping cell buckconfig (by default) to logs
Summary:
Buck dumps all config files for all cells to the logs by default. This creates a huge noise-to-error ratio.
This diff changes the behavior to only dump the buckconfig for the repo (which is checked in anyway) when the logging level
is set to verbose. The behavior for `.buckconfig.local`, `~/.buckconfig.d/` etc remains unchanged.

Test Plan:
Before: P58966187 (~500kb, ~3.4% of information related to actual failure)
After: P58967242 (~150kb, ~10.1% of information related to actual failure)

Reviewed By: philipjameson

fbshipit-source-id: f793348