Fixing deprecated Futures warnings in Buck
Summary: Changed usage of Futures to take an explicit executor. This perserves behavior while removing the deprecated warnings. I did not assess if the exectors should be changed to be more secure per the message. This class of deprecated warnings is different than the ones listed in the Task.

Test Plan:
CI

Also ran unit tests and built two repos

Reviewed By: ttsugriy

fbshipit-source-id: 30e346b