Refactor Main to bring all executors under try-with-resources
Summary:
Executor service is not a closeable resource, but in fact it should be used as closeable pattern - failure to properly shutdown an executor service may lead to hanging threads and application inability to finish. Existing executor service's lifecycle is potentially fragile - executors are created and destroyed in order but do not have appropriate try-finally block each (only one common block, which is defined in wrong scope).
The proposed solution uses try-with-resources on executors (RAII idiom) which is much better:
- proper order of handling of potential errors during executor service's initialization
- better error suppression during executor service shutdown due to try-with-resources exception suppression mechanism

Test Plan: This is pure refactoring so existing unit tests should do. I will create tests for new class CloseableWrapper later if people like the idea.

Reviewed By: bobyangyf

fbshipit-source-id: 59df4fa