Make go propagate dependencies on buildtargets in srcs
Summary: Previously, if a build target was added to srcs, it was not added to the actual dependences of the rule. This makes sure we don't forget that step.

Test Plan: Unit tests. Also, sample project where we sleep for 5 seconds before outputing source worked, where before it would not wait for the job to finish, and just assume that the file was missing.

Reviewed By: styurin

fbshipit-source-id: 052001c