Allow cross-cell library dependencies
Summary:
We were filtering out cross-cell deps from generated buck projects. This caused
missing symbols in IJ. It turns out that while this is necessary for source folders, for
library jars, this is not needed, and we can get better symbol resolution by removing
this filter.

Test Plan:
I tested this manually -- IJ 2017.2 allows library references external to the project,
and with cross-cell the library reference is actually internal to the project, so in both cases it works.
I added a new cross-cell integration test as well.

Reviewed By: mzlee

fbshipit-source-id: 125b789