Introduced CleanBuildShutdownException to kill local build without spamming console
Summary: Current behavior is to display an error stack trace in the console if Build.terminateBuildWithFailure(..) called, where as for Stampede we want to silently shut down the local build and displaying the shutdown Exception is not important.

Test Plan: run some builds and ensure that Exception passed to terminateBuildWithFailure is not displayed.

Reviewed By: shivanker

fbshipit-source-id: 279106c