Expand directories in Go srcs param
Summary:
As discussed in #518 and #560, to compile the output files of a genrule, we need to make a genrule for each output. This could be problematic when the first genrule produces multiple files and their names are not predictable.

This PR makes it possible to put such genrule target in the `srcs` and compile the all generated files.
Closes https://github.com/facebook/buck/pull/1736

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: e5eb2d8