Convert MacroExpander to return Arg instead of String
Summary:
This changes MacroExpander.expand() to return an Arg and changes
extractRuleKeyAppendables to just return that Arg.

All MacroExpander implementations are updated to return Args. For the
most part, that's just a matter of introducing a new Arg implementation
specific to that MacroExpander that holds the rulekey appendables and
then has an appendToCommandLine() implementation that converts those to
the previous result of expand().

Most non-test users use macros through MacroHandler or one of the
MacroArgs and not MacroExpander directly and so are not updated in this
diff. The next diff will update MacroExpander, but the MacroArgs will
still work just fine (appending them to the command line just expands to
the underlying Arg and appends that to the command line).

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: d998204