Treat all containing file cells as having correct cell name format
Summary:
Since all cell names recorded in context always have their prefixes removed
they all seem as if they are incorrect, which is not true. Even if it was true
the warning must have already been issued for the containing file and there is
no point in issuing it again.

Test Plan: CI

Reviewed By: philipjameson

fbshipit-source-id: d32f5a4