pch (10/10): use normal header if precompiling is disabled.
Summary:
NOTE: This is attempt #2.  Previous attempt was reverted due to problem related to pre-includes not happening in Infer-capture-all builds.

If `cxx.pch_enabled` is not true, but `precompiled_header` is used, then don't issue an error.  If `precompiled_header` or `prefix_header` is used, then include the original `.h` file without precompiling.

This removes the "PCH Unavailable Mode" which is now pretty much useless.  Now, if precompiling is unavailable, we'll fail over to a plaintext (non-compiled) header instead.

Test Plan:
CI + local testing with:

  ant clean
  ant && bin/buck build buck && bin/buck test //test/com/facebook/buck/cxx:cxx --filter '.*(Precompi|pch|Pch|PCH|Infer).*'
  arc lint
  ant lint

Reviewed By: dinhviethoa

fbshipit-source-id: 53033f5