Add http_file build rule
Summary: This adds an 'http_file' rule. This is somewhat similar to remote_file, but it's more restricted in its purpose and adds sha256 support and multiple-uri support. Archiving logic will be broken into a different rule down the line. Remote file will not be removed any time soon as people are actively using it right now.

Test Plan: Added tests, ran soyweb for the docs

Reviewed By: ttsugriy

fbshipit-source-id: 14f7b69