Remove custom AddsToRuleKey implementation stuff
Summary:
Currently, MBR buildables compute their own keys via the custom class
info. We require that all fields are final and they are all added to the
rulekeys. This makes it really odd to migrate to MBR where an
RuleKeyAppendable object might be shared across MBR rules and legacy
rules and then its rulekey is computed slightly different in the two
contexts.

This changes MBR to use the standard rulekey computation. It is updated
to require all implementations to have AddToRuleKey specified on all of
their fields.

If we ever get to a state where all rules are MBR-based, I think we will
want to bring back the custom rulekey computation as it should actually
be faster than the current approach.

Test Plan: CI, tests updated

Reviewed By: ttsugriy

fbshipit-source-id: 380e04c