Delete InputPath
Summary:
This changes MBR to allow SourcePath in fields and just treats any
encountered SourcePath as an input. This will make it much easier to
migrate to MBR as previously we couldn't really share referenced
RuleKeyAppendables from both an AbstractBuildRule and a ModernBuildRule
if they had any input paths (as one required it to be a SourcePath and
the other required it to be an InputPath).

InputPath and the support around it is deleted.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 6e15805