Xcode project support testing modular apple_library
Summary:
If we are testing a modular apple_library, we expose it non-modular. This allows the
testing target to see both the public and private interfaces of the tested target
without triggering header errors related to modules. We hide the module definition by
using a filesystem overlay that overrides the module.modulemap with an empty file.

Note: stacked on top of #1561 and #1582.
Closes https://github.com/facebook/buck/pull/1587

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: 8f5e43e