[TravisCI] Do not install unnecessary packages.
Trusty now has a GCC 4.8 and a GHC 7.6 packages available by default,
so there is not need to install them from external sources.

Also, default version of GCC is most likely already pulled in by some
other dependencies, so in the end we should end up pulling in fewer
packages.