Wait for event bus before listeners shutdown
Summary:
Not all event listeners are created equal. SuperConsole event listener is explicitly listed in try-with-resources block and closed before we wait for all events to complete, that leads to race condition - the final super console frame may not be sent to the client. Build finishes but UI does not say so.
This is a temporary solution before we refactor all listeners to follow proper deinitialization cycle.

Test Plan: tests, buck run locally, cancellations

Reviewed By: ttsugriy

fbshipit-source-id: a403be4