Move "Go to Buck file" to Navigate menu
Summary:
Moving "Go to Buck file" to more appropriate places.

Menu action was in the "Build" section which is totally wrong, moved it to "Navigate":

{F120305137}

Popup action was the first action in the popup. Moved it to the "Go to" subsection:

{F120305144}

Also changed these actions not to have icons to be consistent with other actions. The action in the toolbar still have an icon and a slightly different text.

Test Plan: Build and install plugin, make sure it looks like on the screenshots.

Reviewed By: ttsugriy

fbshipit-source-id: 22cfbd5