[TravisCI] Prepare build matrix for OSX support.
Buck has a number of Apple integration tests which are skipped on
Linux OS. Apple test health is extremely important for Buck, so this
change is a preparation for adding OSX to a build matrix.

There are currently test failures on OSX, so this change prepares
the matrix configuration, but does not enable any of CI actions.

This change also speeds up jobs by not removing unnecessary package
installation. For example, unit tests do not require things like GCC
and most android components.