superconsole: Do not halt rendering on dirty stdout
Summary:
Previously we were halting rendering if stdout was dirty in the superconsole. We don't write to stdout save for one place in test results, so drop this requirement. This was done because some commands (like buck audit rules, buck query, etc) write directly to stdout and this can cause errors to stop being written.
buck query got lucky by the order that it parses files in, but buck audit rules does one file at a time, and for the skylark parser (which writes over console events) we wouldn't print out parser errors because '# foo/BUCK' had been written to stdout.
I believe that the locking within DirtyPrintStreamDecorator should be sufficient for the one stdout usage we have in the superconsole event handler.

Test Plan: Updated integration test for buck audit rules, and added a test to super console to make sure this behavior is explicit (and that the test should break if we change it.). CI for the rest.

Reviewed By: ttsugriy

fbshipit-source-id: 46afbed