Work around a bug in Trees.getTypeDecls
Summary:
`Trees.getTypeDecls` will return some `ImportTree`s if there is an `EmptyStatementTree`
in the midst of the imports. (That is, if someone adds a stray semicolon.)

I checked each of our uses of `getTypeDecls` and found only one that was affected --
`TreeBackedEnter` was using that to drive its type-entering loop, and it was throwing
when encountering a non-`ClassTree`. Because it was also using `getTypeDecls` being empty
as a signal to avoid work, I switched to using a `TreePathScanner` to iterate the top-level types
rather than doing so manually.

Test Plan: Existing unit tests, plus a couple new cases

Reviewed By: dreiss

fbshipit-source-id: 0aec57f