Improve error message when failing to load class usage files
Summary:
Changes the wrapped exception to a BuckUncheckedExecutionException to
get the nice context printing from that. Also resolves the full path so
there's no need to know what cell the failure was from.

Test Plan: I hit this error in testing (and it actually threw an NPE because the thrown exception had no localized message). This change gave a much better error.

Reviewed By: styurin

fbshipit-source-id: c2d066b