Improve pipeline precondition error message
Summary:
Adds a separate error message for when a rule doesn't depend on the
previous rule in the pipeline.

Adds more information when the deps contain extra deps not in the
previous rule's deps (added the rule with extra deps and the list of
extra deps).

Test Plan: Added tests.

Reviewed By: ttsugriy

fbshipit-source-id: 539a2e0