Synchronized build phase now uses its own cache fetch thread pool
Summary:
During the synchronized build phase, the majority of work is downloading from the cache, and so this should be less restricted than normal.

Default thread pool size for this phase is now 20, rather than 5 for regular Buck builds or Stampede racing build.

The number of threads to be used can be changed with the following config option:

```
[cache]
sync_build_http_cache_fetch_threads=10
```

Test Plan: ci

Reviewed By: shivanker

fbshipit-source-id: 63ac548