Fix buck killall when Idea is running on Mac
Summary:
Before this patch `buck killall` fails if idea is running.

That's because `jps -l` doesn't output main class name for Idea
process.  This is likely caused by the fact that Idea is launched
not as `java Main`, but it is native application with embeddeded
JVM.

Test Plan:
```
% buck killall
2018-02-14 13:21:16,393 [ERROR][buck.py:115] cannot parse a line in jps -l outout: '74626 \n'
Traceback (most recent call last):
  File "/Users/nga/.dewey/buck/.commits/fea4ff23437db25d23da3e7794f2dad4d32f9d46/cli/macosx/buck/programs/buck.py", line 89, in <module>
  File "/Users/nga/.dewey/buck/.commits/fea4ff23437db25d23da3e7794f2dad4d32f9d46/cli/macosx/buck/programs/buck.py", line 56, in main
  File "/Users/nga/.dewey/buck/.commits/fea4ff23437db25d23da3e7794f2dad4d32f9d46/cli/macosx/buck/programs/buck.py", line 33, in killall_buck
Exception: cannot parse a line in jps -l outout: '74626 \n'
% ps aux | grep -i buck | grep -v grep | wc -l
1
% ./buck-out/gen/programs/buck.pex killall
% ps aux | grep -i buck | grep -v grep | wc -l
0
```

Reviewed By: dinhviethoa

fbshipit-source-id: a52e0c1