Revert superconsole change
Summary: This reverts a change to the superconsole with respect to checking for stdout dirtiness. Without this fix, some output would be written, but it would appear to be overwritten in an interactive shell when we cleared the stderr lines.

Test Plan: buck audit rules is appropriately broken for skylark parser (will be fixed in another diff), and buck build --show-outputs shows outputs in an interactive shell again.

Reviewed By: ttsugriy

fbshipit-source-id: ff1109b