Extract prebuilt python archives
Summary: This makes prebuilt_python_library extract whls and eggs to a location off to the side. This is done so that in a future diff we'll be able to link all of the sources into one larger directory, ideally leading to better performance with inplace binaries.

Test Plan: Updated integration tests

Reviewed By: andrewjcg

fbshipit-source-id: 5bfe671