Incremental action graph construction for a subset of build rules deemed safe
Summary:
Adds support for incremental action graph construction for a subset of build rules (mostly C++/Python). Disabled by default. Sorry for the large diff, but I think this is easier to understand as a whole.

The proper long term solution is likely something similar to Bazel's provider model. Unfortunately, given our current architecture, there is some complexity involved in supporting incrementality in the short term.

`ActionGraphCache` now uses `ActionGraphNodeCache` to cache `BuildRule` subgraphs for `TargetNode`s. The bottom-up traversal of the target graph in `ActionGraphCache` short circuits when it hits a `TargetNode` whose subgraph is cached, avoiding reconstruction of that subgraph.

When we detect that a `TargetNode` has changed, we invalidate its parent chains in the cache. Similarly, if we find a `TargetNode` whose `BuildRule` isn't cacheable, to be safe we also invalidate its parent chains.

`BuildRule`s that participate in caching must implement `CacheableBuildRule`, so we know they don't violate our assumptions about touching the action graph outside of the scope of the parent `TargetNode`. This interface also allows a `BuildRule` to expose additional "implicit" dependencies for caching which are not exposed as part of `getBuildDeps` and `getRuntimeDeps`. `CxxLibrary` uses this to expose its associated `Archive` node for caching, which is not actually in `CxxLibrary`'s subgraph, but needs to be cached to avoid the cost of recomputing that node when constructing the parent `CxxBinary`.

Unfortunately, many `BuildRule`s hold on to `BuildRuleResolver`s and related objects, and removing this state would involve significant refactoring. To allow cached `BuildRule`s to rely on up-to-date `BuildRuleResolver`s during action graph construction (parents that we need to recreate will sometimes call into cached children), and to prevent leaking previous action graphs, `CacheableBuildRule` also includes an `updateBuildRuleResolver` method, allowing `BuildRule`s to update their state when a new action graph is built incrementally.

Still remaining (change was already getting too large):
  1. Add instrumentation
  2. Add deps caching/fix some todos
  3. Add C++/Python integration tests
  4. See if we can speed this up even further (slowed down a bit from my earlier experiments)

Test Plan: Added unit tests + ad hoc testing.

Reviewed By: bobyangyf

fbshipit-source-id: 796ee93