Replace Parser with DaemonicParserState in PerBuildState
Summary:
This change removes the dependency of `PerBuildState` on `Parser` by passing `DaemonicParserState` directly.

It seems that `DaemonicParserState` was designed initially to be an implementation detail of `Parser`, but it's accessed from `PerBuildState`.

The state after these changes is not ideal, but it's better to not have cyclic dependencies between classes.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 0b4517c