Don't be so restrictive of static fields reachable from Buildables
Summary:
Previously we were creating FieldInfos (such that deps, inputs, and
other derivations would consider them) for static fields and
requiring that they were "simple". This is unnnecessarily restrictive
and makes it difficult to write tests (as any Buildables defined in the
tests will have access to statics of the test).

Remove the the "simple" restriction and the creation of a FieldInfo, but
keep the requirement that the fields are final. This should be
sufficient that these fields don't reference things that shouldn't be
accessible from the Buildable.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 2feb5e6