Introduce ValueVisitor
Summary:
This allows deriving values from the fields of a Buildable by defining a
Visitor (instead of having to implement a method in all of the different
ValueTypeInfo implementations).

Introduce a simple implementation of ValueVisitor that generates a
string representation of the Buildable. This isn't really intended to be
used for anything.

A followup will convert the current derived values to be done via
visitors.

Test Plan: Added tests for the stringifying visitor

Reviewed By: ttsugriy

fbshipit-source-id: cce5d32