Add support for fields that implement AddsToRuleKey
Summary:
When we encounter a value that doesn't have a more specific
ValueTypeInfo, if that value implements AddsToRuleKey we return
DynamicTypeInfo.INSTANCE.

When visiting such a value, we will look up the ClassInfo for the
runtime type of the object and delegate to its visit. In this way, we
will traverse its fields in turn.

Test Plan: CI, added tests

Reviewed By: ttsugriy

fbshipit-source-id: 48e776d