Add BuildTargetValueTypeInfo
Summary:
This adds support for BuildTarget fields. This will make it much easier
to support serialization of BuildTargetSourcePaths (and some rules
actually have BuildTarget fields).

This is the first instance of a ValueTypeInfo that doesn't have a
corresponding method on ValueTypeVisitor and instead makes multiple
calls into the visitor for its `visit()` implementation. It's possible
that we should add something explicit on the visitor to indicate that
it's visiting some sort of composite object, but that's not necessary
yet.

Test Plan: Added tests for all ValueVisitors

Reviewed By: ttsugriy

fbshipit-source-id: 44bd599