tone down cache error messages
Summary:
The messages were over-alarming. I toned them down so they are printed in the cli only after we report 100 of those but we still log every single one of them in the log.

I also added documentation for the buckconfig option.

Test Plan: tweaked the necessary tests to see that it works.

Reviewed By: ruibm

fbshipit-source-id: da45ac4