Generalize custom compiler usage in ant.
Summary:
Current approach hard-codes ECJ compiler and makes which
makes it hard to experiment with other compilers.

With this change, in order to change a compiler all it takes is to
change the value of `buck.build.compiler` and a `buck.build.compiler.path`.

This is in preparation for enabling a way to use error-prone compiler.

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: 9617b63