Do not use allKnownRules from the RuleKeyCalculator to fix race.
Summary: We need to make a MultiContains request upfront for pruning optimally. Earlier, the keys for this were bing obtained using `ParallelRuleKeyCalculator::getAllKnownRules`, but we were not waiting for the top-level rule-key calculation to finish. A simplistic way to fix this race is to just use the list of rules from first pass over the graph, which we anyway need for other reasons.

Test Plan: CI, tested on personal tier.

Reviewed By: alisdair04

fbshipit-source-id: 4461352