Incremental action graph: cache deps
Summary: Adds deps caching for incremental action graph generation. Note that the performance benefit of this is actually minimal. This is needed for correctness in the scenario in which a `BuildRule` stays in our cache for a few action graph constructions, but doesn't get used by `requireRule`. In this case, we'd previously lose the ability to compute that rule's runtime deps, as we wouldn't have access to the `BuildRuleResolver` that rule was originally generated with, and we can't cache the original `BuildRuleResolver` for each rule in the cache since we'd leak a bunch of action graphs.

Test Plan: Added unit test coverage + adhoc testing.

Reviewed By: bobyangyf

fbshipit-source-id: c8ba998