Add serialization/deserialization
Summary:
Adds a ValueCreator interface that's very similar to ValueVisitor and
then adds the corresponding create() methods similar to visit() methods.

Serialization/Deserialization is then basically just implemented with a
ValueVisitor/ValueCreator combination that writes to/reads from a
DataInput/OutputStream.

Referenced "dynamic" objects (i.e. implementations of AddsToRuleKey) are
serialized as simply a hash of their serialized representation. The
serialization of a complex object is then effectively a merkle tree. This
allows us to share the serialized representation of shared objects (for c++
particularly, there are many shared references to the PreprocessorDelegate and
other such fields).

Deserialization uses Objenesis to instantiate objects and then directly sets
the fields from the serialized form of the fields. This depends on all of the
fields being serialized, but the MBR framework currently requires that anyway
(I plan to weaken this in the future and then will need to deal with how it
interacts with serialization).

Test Plan:
Added an implementation of AbstractValueVisitorTest that
serializes/deserializes all the types from that abstract and verifies that the
stringified versions of the before and after match.

Reviewed By: ttsugriy

fbshipit-source-id: ce6d7f6