Cancel Stampede FileHashComputation as soon as a build gets rejected.
Summary: Async BuildJobState computation depends on async FileHashes computation. Both of them need to be cancelled when stampede build fails (or is cancelled) prematurely. Otherwise we pay the cost of computing file hashes unnecessarily. This diff chains the 2 futures such that when `asyncJobState` is cancelled, it cancels `distBuildFileHashes` whether or not it started to run in the first place.

Test Plan: Manual

Reviewed By: alisdair04, michsien

fbshipit-source-id: e0e81bb