Avoid some exceptions in getAnnotation
Summary:
When underlyingConstruct.getAnnotation is called on an annotation that
includes an array of classes, where one of the classes is not available,
javac throws a nearly-useless ClassCastException.

Make getAnnotation (try to) return an annotation object that we
construct instead.  It is a dynamic proxy that tries to use an
AnnotationValueVisitor to handle requests for fields.  If it can't, it
falls back to calling getAnnotation on the underlying construct.

This allows APs that use getAnnotation to work even if they contain
arrays of unavailable classes, as long as they only try to extract
simple data objects from the annotations.

Test Plan: Added tests.

Reviewed By: jkeljo

fbshipit-source-id: af16230