Provide a better exception when failing to get annotation fields
Summary:
Currently, this is failing with an unhelpful exception when an
annotation argument is an array of classes, and one of the class
literals is not present.

It's not clear that we can turn this into a normal compile error without
returning some value that would confuse the annotation processor.
However, we can at least point to the problematic element, which will
help with migration.

Test Plan: Updated test.

Reviewed By: jkeljo

fbshipit-source-id: 536fe2c