Add support for @classpath_abi macro
Summary:
This is similar to classpath macro but gives the corresponding
class-abi jar for the classpath entries.

This is useful for genrules which only depends on the dependencies abi’s.
For eg: android lint where it need not run the lint checks again if
the dependent’s abi doesn’t change.

Usage: `$(classpath_abi :<target_path>)`
Closes https://github.com/facebook/buck/pull/1780

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: ebea70b