Do not require platform deps for non requested platform.
Summary:
Previously, all platform_deps would be added to extraDeps
and as such would be built even if target platform is actually different.
It was always broken, but not noticeable when libraries are specified
as platform deps, since their default flavors result in essentially noop
rules, but in case something like `apple_bundle` is added as a dependency,
it would actually be built, which is not desired.

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: acc68d5