Fixing Machos scrubbing methods to handle cgo files
The OSO Symbols being sanitized in Mach-O files are found in segments under LC_SEGMENT(_64) load commands named "__LINKEDIT". These were being identified by finding LC_SEGMENT(_64) commands that mapped to data at the end of the binary file. This was a hacky approach. Also, these segments are not present in cgo binaries, so the scrubber was erroneously failing when there was actually nothing to scrub. Now if there are no segments named "__LINKEDIT" or LC_UUID commands in the binaries, the scrubber returns successfully (as there was nothing to scrub).